refactor #9 Move get_independencies to DAG
Moves get_independencies from BayesianModel to DAG
